pyproject.toml
[project]
name = "staffai"
version = "0.1.0"
description = "一人公司本地专用 AI 员工系统"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=1.0.0",
    "gradio>=5.0.0",
    "pyyaml>=6.0",
]
